Y3ZKM2D ask git whether a rebase is in progress, so a failed abort is not read as nothing to abort
